# docs/api/openapi/simulation-engine.yaml
openapi: 3.1.0
info:
  title: Simulation Engine
  version: 0.1.0
paths:
...
  /api/v1/aas/shells:
    get:
      summary: List Shells
      operationId: list_shells_api_v1_aas_shells_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      summary: Create Shell
      operationId: create_shell_api_v1_aas_shells_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AasCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/aas/submodels:
    post:
      summary: Create Submodel
      operationId: create_submodel_api_v1_aas_submodels_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AasSubmodelCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/aas/submodels/{submodel_id}/elements:
    get:
      summary: Get Submodel Elements
      operationId: get_submodel_elements_api_v1_aas_submodels__submodel_id__elements_get
      parameters:
      - name: submodel_id
        in: path
        required: true
        schema:
          type: string
          title: Submodel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Patch Submodel Elements
      operationId: patch_submodel_elements_api_v1_aas_submodels__submodel_id__elements_patch
      parameters:
      - name: submodel_id
        in: path
        required: true
        schema:
          type: string
          title: Submodel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AasSubmodelElementsPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/aasx/upload:
    post:
      summary: Upload Aasx
      operationId: upload_aasx_api_v1_aasx_upload_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AasxUploadRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AasCreate:
      properties:
        aas_identifier:
          type: string
          title: Aas Identifier
        product_identifier:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Identifier
        product_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Name
        product_category:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Category
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      type: object
      required:
      - aas_identifier
      title: AasCreate
    AasSubmodelCreate:
      properties:
        submodel:
          type: object
          title: Submodel
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
        register:
          type: boolean
          title: Register
          default: true
      type: object
      required:
      - submodel
      title: AasSubmodelCreate
    AasSubmodelElementsPatch:
      properties:
        elements:
          title: Elements
      type: object
      required:
      - elements
      title: AasSubmodelElementsPatch
...
    AasxUploadRequest:
      properties:
        filename:
          type: string
          title: Filename
        content_base64:
          type: string
          title: Content Base64
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      type: object
      required:
      - filename
      - content_base64
      title: AasxUploadRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError